requirements.txt
django
djangorestframework
requests
django-cors-headers
djangorestframework-simplejwt